+++ CHANGES.md
@@ -144,6 +144,13 @@
     - Fixed an issue where the bottom right window would not respond to
       `MirrorShrink` and `MirrorExpand` messages.
 
+  * `XMonad.Hooks.EwmhDesktops`
+
+    - Added `disableEwmhManageDesktopViewport` to avoid setting the
+      `_NET_DESKTOP_VIEWPORT` property, as it can lead to issues with
+      some status bars (see this
+      [polybar issue](https://github.com/polybar/polybar/issues/2603)).
+
 ### Other changes
 
 ## 0.17.1 (September 3, 2022)